# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduCycle API
  description: |
    Academic Resource Exchange Platform API
    
    This API allows students to exchange academic resources such as books and notes.
    Features include location-based discovery, institution-based filtering, and secure authentication.
    
    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - General endpoints: 100 requests per minute
    - Auth endpoints: 5 requests per 15 minutes
    - Search endpoints: 30 requests per minute
    
    ## Base URL
    - Development: `http://localhost:3001/api`
    - Production: `https://api.educycle.com/api`
  version: 1.0.0
  contact:
    name: EduCycle Support
    email: support@educycle.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.educycle.com/api
    description: Production server
...
paths:
...
  /listings:
    post:
      tags:
        - Listings
      summary: Create listing
      description: Create a new book or note listing
      operationId: createListing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListing'
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /listings/{id}:
    get:
      tags:
        - Listings
      summary: Get listing by ID
      description: |
        Get detailed listing information. By default returns a flat listing (no nested user).
        Set **include=user** to include the listing owner (id, name, email, phone, verified) in the response.
      operationId: getListingById
      parameters:
        - $ref: '#/components/parameters/ListingId'
        - name: include
          in: query
          required: false
          schema:
            type: string
            enum: [user]
          description: Include nested user object when set to "user"
      responses:
        '200':
          description: Listing details (flat or with user when include=user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                flat:
                  summary: Without include
                  value:
                    success: true
                    data:
                      id: "uuid"
                      user_id: "uuid"
                      title: "Class 10 Mathematics"
                      category_code: "school_9_10"
                      type: "sell"
                      price: 200
                withUser:
                  summary: With include=user
                  value:
                    success: true
                    data:
                      id: "uuid"
                      user_id: "uuid"
                      title: "Class 10 Mathematics"
                      user:
                        id: "uuid"
                        name: "John Doe"
                        email: "john@example.com"
                        verified: true
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags:
        - Listings
      summary: Update listing
      description: Update own listing (only owner can update)
      operationId: updateListing
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListing'
      responses:
        '200':
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - Listings
      summary: Delete listing
      description: Delete own listing (only owner can delete)
      operationId: deleteListing
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /listings/search:
    get:
      tags:
        - Listings
      summary: Search listings
      description: Search listings with filters and location
      operationId: searchListings
      parameters:
        - name: categoryCode
          in: query
          schema:
            $ref: '#/components/schemas/CategoryCode'
        - name: type
          in: query
          schema:
            type: string
            enum: [sell, rent, free]
        - name: condition
          in: query
          schema:
            type: string
            enum: [new, like_new, good, fair, poor]
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: number
            minimum: 0
        - name: freeOnly
          in: query
          schema:
            type: boolean
        - name: radius
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Radius in kilometers (requires latitude/longitude)
        - name: latitude
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: institutionId
          in: query
          schema:
            type: string
            format: uuid
        - name: query
          in: query
          schema:
            type: string
            maxLength: 255
          description: Full-text search query
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursorId
          in: query
          schema:
            type: string
            format: uuid
          description: Listing UUID for cursor (use with cursorCreatedAt for next page)
        - name: cursorCreatedAt
          in: query
          schema:
            type: string
            format: date-time
          description: ISO timestamp of last item for next page
        - name: includeTotal
          in: query
          schema:
            type: boolean
            default: false
          description: Include total matching count (extra query; use sparingly)
      responses:
        '200':
          description: Search results (cursor-based pagination)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
  /listings/me:
    get:
      tags:
        - Listings
      summary: Get my listings
      description: Get authenticated user's listings
      operationId: getMyListings
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: User's listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedListings'
...
  /upload/listings:
    post:
      tags:
        - Upload
      summary: Upload single listing image
      description: Upload one image to S3. Resized and thumbnail generated. Returns url and thumbnailUrl for use in listing images array. Max 10MB; JPEG/PNG/WebP.
      operationId: uploadListingImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
                      thumbnailUrl:
                        type: string
                        format: uri
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
...
    ListingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Listing UUID
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    CategoryCode:
      type: string
      enum:
        - school_1_5
        - school_6_8
        - school_9_10
        - school_11_12_science
        - school_11_12_commerce
        - school_11_12_arts
        - college_ba
        - college_bsc
        - college_bcom
        - college_bba
        - college_bca
        - college_btech
        - college_mba
        - college_law
        - college_medical
        - college_pharmacy
        - competitive_upsc
        - competitive_ssc
        - competitive_banking
        - competitive_railways
        - competitive_defence
        - competitive_state
    CreateListing:
      type: object
      required:
        - title
        - categoryCode
        - type
        - images
        - latitude
        - longitude
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        categoryCode:
          $ref: '#/components/schemas/CategoryCode'
        type:
          type: string
          enum: [sell, rent, free]
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        price:
          type: number
          minimum: 0
          maximum: 1000000
          description: Required for sell/rent, not allowed for free
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 5
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        institutionId:
          type: string
          format: uuid
    UpdateListing:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        categoryCode:
          $ref: '#/components/schemas/CategoryCode'
        type:
          type: string
          enum: [sell, rent, free]
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        price:
          type: number
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
          maxItems: 5
        latitude:
          type: number
        longitude:
          type: number
        isActive:
          type: boolean
...
    SearchResult:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                listings:
                  type: array
                  items:
                    $ref: '#/components/schemas/Listing'
                limit:
                  type: integer
                hasMore:
                  type: boolean
                nextCursor:
                  type: object
                  nullable: true
                  properties:
                    cursorId:
                      type: string
                      format: uuid
                    cursorCreatedAt:
                      type: string
                      format: date-time
                total:
                  type: integer
                  description: Present only when includeTotal=true
    PaginatedListings:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                data:
                  type: array
                  items:
                    $ref: '#/components/schemas/Listing'
                pagination:
                  type: object
                  properties:
                    page:
                      type: integer
                    limit:
                      type: integer
                    total:
                      type: integer
                    totalPages:
                      type: integer
                    hasMore:
                      type: boolean
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - path: email
                  message: Invalid email address
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have permission to access this resource
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...